requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
Markdown==3.7
icalendar==6.1.1
recurring-ical-events==3.8.0
requests==2.32.3
python-dotenv==1.0.1
